# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Langinfra",
    "version": "1.3.4"
  },
  "paths": {
...
    "/api/v1/upload/{flow_id}": {
      "post": {
        "tags": [
          "Base"
        ],
        "summary": "Create Upload File",
        "description": "Upload a file for a specific flow (Deprecated).\n\nThis endpoint is deprecated and will be removed in a future version.",
        "operationId": "create_upload_file_api_v1_upload__flow_id__post",
        "deprecated": true,
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_upload_file_api_v1_upload__flow_id__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/langinfra__api__v1__schemas__UploadFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/upload/{flow_id}": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_api_v1_files_upload__flow_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_files_upload__flow_id__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/langinfra__api__v1__schemas__UploadFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/download/{flow_id}/{file_name}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Download File",
        "operationId": "download_file_api_v1_files_download__flow_id___file_name__get",
        "parameters": [
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Name"
            }
          },
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/images/{flow_id}/{file_name}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Download Image",
        "operationId": "download_image_api_v1_files_images__flow_id___file_name__get",
        "parameters": [
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Name"
            }
          },
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/profile_pictures/{folder_name}/{file_name}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "Download Profile Picture",
        "operationId": "download_profile_picture_api_v1_files_profile_pictures__folder_name___file_name__get",
        "parameters": [
          {
            "name": "folder_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Folder Name"
            }
          },
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/files/list/{flow_id}": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List Files",
        "operationId": "list_files_api_v1_files_list__flow_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/delete/{flow_id}/{file_name}": {
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete File",
        "operationId": "delete_file_api_v1_files_delete__flow_id___file_name__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ],
        "parameters": [
          {
            "name": "file_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Name"
            }
          },
          {
            "name": "flow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/files/": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List Files",
        "description": "List the files available to the current user.",
        "operationId": "list_files_api_v2_files__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/langinfra__services__database__models__file__model__File"
                  },
                  "type": "array",
                  "title": "Response List Files Api V2 Files  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      },
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload User File",
        "description": "Upload a file for the current user and track it in the database.",
        "operationId": "upload_user_file_api_v2_files__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_user_file_api_v2_files__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/langinfra__api__schemas__UploadFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete All Files",
        "description": "Delete all files for the current user.",
        "operationId": "delete_all_files_api_v2_files__delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      }
    },
    "/api/v2/files": {
      "get": {
        "tags": [
          "Files"
        ],
        "summary": "List Files",
        "description": "List the files available to the current user.",
        "operationId": "list_files_api_v2_files_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/langinfra__services__database__models__file__model__File"
                  },
                  "type": "array",
                  "title": "Response List Files Api V2 Files Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      },
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Upload User File",
        "description": "Upload a file for the current user and track it in the database.",
        "operationId": "upload_user_file_api_v2_files_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_user_file_api_v2_files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/langinfra__api__schemas__UploadFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "summary": "Delete All Files",
        "description": "Delete all files for the current user.",
        "operationId": "delete_all_files_api_v2_files_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "API key query": []
          },
          {
            "API key header": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_upload_file_api_v1_upload__flow_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_upload_file_api_v1_upload__flow_id__post"
      },
...
      "Body_upload_file_api_v1_files_upload__flow_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_files_upload__flow_id__post"
      },
...
      "Body_upload_user_file_api_v2_files__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_user_file_api_v2_files__post"
      },
      "Body_upload_user_file_api_v2_files_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_user_file_api_v2_files_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "langinfra__api__schemas__UploadFileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "size"
        ],
        "title": "UploadFileResponse",
        "description": "File upload response schema."
      },
      "langinfra__api__v1__schemas__UploadFileResponse": {
        "properties": {
          "flowId": {
            "type": "string",
            "title": "Flowid"
          },
          "file_path": {
            "type": "string",
            "format": "path",
            "title": "File Path"
          }
        },
        "type": "object",
        "required": [
          "flowId",
          "file_path"
        ],
        "title": "UploadFileResponse",
        "description": "Upload file response schema."
      },
      "langinfra__services__database__models__file__model__File": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "path",
          "size"
        ],
        "title": "File"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/login"
          }
        }
      },
      "API key query": {
        "type": "apiKey",
        "in": "query",
        "name": "x-api-key"
      },
      "API key header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}